CLN: refactored SubprocessWorkflowEvaluator class and unit tests into seperate files
